update gemini chatbot
