--- gemini.py
@@ -1,20 +1,0 @@
-import google.generativeai as ai
-
-# API Key
-API_KEY = 'API KEY'
-
-# Configure the API
-ai.configure(api_key=API_KEY)
-
-# Create a new model
-model = ai.GenerativeModel("gemini-pro")
-chat = model.start_chat()
-
-# Start a conversation
-while True:
-    message = input('You: ')
-    if message.lower() == 'bye':
-        print('Chatbot: Goodbye!')
-        break
-    response = chat.send_message(message)
-    print('Chatbot:', response.text)
